workflow/create_snapshot_and_image/requirements.txt
six
google-cloud-compute
google-cloud-storage
google-cloud-build
